requirements.txt
pyjwt
Async-OAuthlib
quart
hypercorn
disnake
PyYAML